Extrayendo el service del runner
